Move Training- and Trainee-related views, forms, filters
The new location: `trainings` application.
